+++ app/ui/src/main/kotlin/org/pcsoft/app/aighost/app/Launcher.kt
@@ -15,14 +15,20 @@ package org.pcsoft.app.aighost.app
 import de.saxsys.mvvmfx.MvvmFX
 import javafx.application.Application
 import javafx.stage.Stage
+import org.apache.commons.lang3.SystemUtils
 import org.pcsoft.app.aighost.app.ui.window.MainWindow
+import org.pcsoft.app.aighost.app.util.PreferencesLoader
+import org.pcsoft.app.aighost.app.util.logger
 
 /**
  * JavaFX application entry point of AI Ghost.
  */
 class AiGhostApplication : Application() {
     override fun start(stage: Stage) {
+        PreferencesLoader.loadPreferences(this)
+
         AiGhostTheme.install()
+
         MainWindow().show()
     }
 }
@@ -31,6 +37,13 @@ class AiGhostApplication : Application() {
  * Starts the application after registering the global resource bundle used for I18N.
  */
 fun main() {
+    logger<AiGhostApplication>().apply {
+        info("====================")
+        info("Starting application")
+        info("System: {}, Java: [Version: {}, Vendor: {}]", SystemUtils.OS_NAME, SystemUtils.JAVA_VERSION, SystemUtils.JAVA_VENDOR)
+        info("====================")
+    }
+
     MvvmFX.setGlobalResourceBundle(Messages.bundle)
     Application.launch(AiGhostApplication::class.java)
 }